Property for the true remote address of a socket
In newer versions of Node.js (or at least 0.8.3 that I tested on),
socket.remoteAddress is replaced by the address in the
X-Forwarded-For header, if it exists. This header can be spoofed.

This adds socket.trueRemoteAddress, for the true remote address of
the socket that won't be replaced by a spoofed IP.

For instance, if a user with IP 192.168.0.101 initiates a connection
with the header X-Forwarded-For: 8.8.8.8, socket.remoteAddress will
be '8.8.8.8' and socket.trueRemoteAddress will be '192.168.0.101'.